:recycle: #28 QiitaのUserID入力FormをHCOで定義するように変更
ついでに名前をQiitaUserFormに変更
